+++ CHANGELOG.md
@@ -132,6 +132,8 @@ call `Function#lock` first in the chain to achieve the same effect.
 - Removed `Object.defaults`. Use `Object.assign` or object spread syntax here
     instead.
 - Removed `Object.values`. Use native ES6 `Object.values` instead.
+- Changed `Object.invert` to not support `multi` second argument. `Object.reduce`
+    can be used here for special use cases such as this.
 
 TODO: